fs: heap/bss calculation ignore section above STACK_TOP.
Today bss/heap is placed above last loaded sections.
Current check is only against TASK_SIZE, maybe since most architectures do:
 define STACK_TOP       TASK_SIZE
But this is not true for all of them and a gap between TASK_SIZE
and STACK_TOP is possible and may be used for variant purposes.
For example we use it for a special physical memory mapping.

In such case of section in this gap we skip it for matter of
heap/bss calculation.
set_brk will never work for such case anyway.

Signed-off-by: Noam Camus <noamc@ezchip.com>