ARC: [plat-eznps] Use dedicated atomic64
This is preperation for our PERF implementation.
So atomic64 ops will be protected from overflow exception as well.

This specific commit is to maintain bisect ability.
All it does is copy the generic implementation - which the next commits will modify.
Then it will be very obvious what are the changes that were done on top of the generic.

Signed-off-by: Noam Camus <noamc@ezchip.com>