ARC: arch_spinlock use cpu_relax() for optimization
In SMT machine with multiple HW threads if one HW thread is
contending on spinlock we wish to switch to another thread
that may hold this spinlock.

This may achieved by using cpu_relax() that should be defined
properly for such platform with SMT.

Running hackbench on NPS400 we got more than 10% improvement in time.

Signed-off-by: Noam Camus <noamc@ezchip.com>