ARC: add cmpxchg64 for CONFIG_HAVE_VIRT_CPU_ACCOUNTING_GEN
This is to ovecome build failure due to commit:
sched, time: Fix build error with 64 bit cputime_t on 32 bit systems
347abad981c1ef815ea5ba861adba6a8c6aa1580

This commit introduce the use of cmpxchg_cputime().
Which in our case defined in asm-generic/cputime_nsecs.h to be:

However ARC as 32 bit system lake to have cmpxchg64.

Signed-off-by: Noam Camus <noamc@ezchip.com>