ARC: [plat-eznps] set possible cpus as present cpus optimization
This optimization is to speed up boot time when 4K CPUs may take
long time on slow platforms such simulator or HW emulator.

This way we are not trying to initialize and boot all 4K CPUs
and we can save a lot of time.
This behaviour is usefull for development and testing.
The trigger is coming from DTS as new option we call
"init-possible-as-present".

Signed-off-by: Noam Camus <noamc@ezchip.com>