ARC: [plat-eznps] add nop before sched instruction
Due to HW bug when code branch to pc of sched instruction
then HW scheduler is doing partial context switch.
This causes reg file of new thread to have content of old
thread for some of the registers.

So at any place we use CTOP_INST_SCHD_RW,CTOP_INST_SCHD_RD and
there is a chance it will be destination of a branch/jump a nop
instruction was prefixed the sched command.

Signed-off-by: Noam Camus <noamca@mellanox.com>