ARC: [plat-eznps] add nop before flag instruction
There is a HW bug that when userspace code uses schd instruction
and the next running thread starts at a flag instruction,
The flag instruction will behave that it is still in userspace
and status32 will not to be updated properly.
A nop instruction before the flag instruction avoids this BUG.

This HW bug can also happen when a HW schedule happens right after
the flag instruction, so what we do is to run the flag instruction
twice. This will ensure us that even if there was a HW schd event
when the first flag instruction was executed, the second flag
instruction will surely update status32 with the correct value.

Signed-off-by: Noam Camus <noamca@mellanox.com>