ARC: [plat-eznps] Add general information after machine check exception.
When the kernel or a data plane application executes an illegal memory
transaction, a machine check exception occurrs. In this case,
information regarding the last transaction or the error cause
is stored in the err_cap_2, err_cap_1 and err_sts registers in
the CIU block.
There's a need to parse the information there in case of a machine
check, and print it to the user.
The following commit provides general information about the
exception, e.g. last transaction of the core, msid, offset in it
and error cause of the exception.

Signed-off-by: Liav Rehana <liavr@mellanox.com>