ARC: [plat-eznps] Add core memory invalid access kernel prints.
Virtual addresses 0x57f00000 to 0x57ffffff of a data plane
application are mapped to CMEM (core memory), where the
section 0x57f00000-0x57f07fff is mapped to the private CMEM
section of the accessing cpu, and the section
0x57f08000-0x57ffffff is mapped to the shared section of the
CMEM that belongs to the accessing core.
When an access exceeds the size of the private or shared sections
of the CMEM, a tlb miss exception occurs and a SIGSEGV is sent to
the application.
There's a need to provide more information as to why this exception
occurred, to the application that caused it.
The following commit adds printings by the kernel that show the sizes
of the different sections of the CMEM (i.e. private, shared and data
cache), and the reason of the exception (exceeding private section,
etc...).

Signed-off-by: Liav Rehana <liavr@mellanox.com>
Signed-off-by: Noam Camus <noamca@mellanox.com>