ARC: [plat-eznps] Add kernel logs regarding fixed mapping table.
Virtual addresses 0x58000000-0x5fffffff of a data plane application
are mapped to 16 FMT (fixed mapping table) slots of equal sizes.
When a data plane application writes to a memory space through an
FMT slot, which has a write privilege (e.g. memory spaces for code
sections in IMEM), the access causes a tlb miss exception.
The following commit adds printings of information about the accessed
memory space and whether or not the FMT slot has write privilege
on cases of tlb miss exception when accessing a virtual address
mapped to an FMT slot.

Signed-off-by: Liav Rehana <liavr@mellanox.com>