ARC: [plat-eznps] Add memory space invalid access kernel prints.
When accessing an invalid (not enabled) memory space or an offset
of a memory space, that is bigger than its size, a machine check
exception occurs and the core that performed the instruction is
halted.
The following commit improves the parsing of the information from
the CIU hardware block, prints more information about the type of
the accessed memory space, the correct offset of the access and
whether or not the memory space is enabled or has write privilege.

Signed-off-by: Liav Rehana <liavr@mellanox.com>