ARC: [DIRTY] Remove define of PTE_SPECIAL for EZNPS platform.
The following commit removes a define that was introduced in the ARC
architecture in kernel version 4.7. That define causes a different
flow of program in the function vm_normal_page under mm/memory.c,
where the check of whether the pfn variable exceeded the
highest_memmap_pfn variable is performed and causes a stack trace
print to the console.
One can reproduce the case which causes a stack trace by building
a simple data plane application for debug and running it.
(e.g. building hello_packet for debug and running it on cpu 16)
The solution here needs to be understanding why the pfn variable
got a higher value than the highest_memmap_pfn variable in the
first place, and then fixing it.

Signed-off-by: Liav Rehana <liavr@mellanox.com>