ARC: [plat-eznps] Ensure no two dp processes run on the same cpu.
There's a need to make sure that no two data plane processes run
on the same cpu.
The following commit ensures the above through the use of a per
cpu variable called dp_running_pid, which is set to the pid of
a process that runs a data plane application and is reset after
the process' death.

Signed-off-by: Liav Rehana <liavr@mellanox.com>