ARC: [plat-eznps] Add dpl ioctl options for access to pct aux regs.
The dpl module provides access for dp applications to auxiliary registers.
Though the registers that are valid for access through this module are
only registers from the Extensions Short Range Allocation space, there is
a need to support access for the 5 pct auxiliary registers in the ARC
Short Range Allocation space.
The following commit provides that support through additional ioctl options
for the dpl module.

Signed-off-by: Liav Rehana <liavr@mellanox.com>