ARC: [plat-eznps] Update GIM config to handle watchdog interrupt.
It was decided, since the NPS watchdog reset doesn't work well with the
SAR values, that the watchdog should send an interrupt outside the NPS
and it will be reset by the host.
The following commit performs the needed configuration to the GIM block,
so it will indeed pass the watchdog interrupt to the int_out line.

Signed-off-by: Liav Rehana <liavr@mellanox.com>