ARC: [plat-eznps] Fix mask/unmask of network or serial irq on cpu other than 0.
In the NPS400 platform of the ARC architecture, cpu 0 handles all the
interrupts from the network or serial interrupts the chip receives.
The ARC architecture provides a special auxiliary register named
IENABLE for masking and unmasking of interrupts. The context of this
register is per cpu. Sometimes, the process that performs the
masking/unmasking of those interrupts is doing so from a cpu other
than 0. In that case, the IENABLE which belongs to cpu 0 is unaffected,
which may leave it masked for the rest of the OS operation, and no
network nor serial interrupts will be received in the chip, making
it appear to be stuck.
The following commit adds a different handling for the network and serial
interrupts, so that the masking/unmasking of those interrupts will be
done only from cpu 0.

Signed-off-by: Liav Rehana <liavr@mellanox.com>