ARC: [plat-eznps] Set nohz_full cpus as housekeeping.
When using a sleep system call, the cpu needs to order a timer to wake
up. Kernel 4.7 introduced another restrict for the cpu that orders the
timer. That cpu needs to be a housekeeping cpu, and that mask is
initialized with all the cpus that are possible in the system and not
nohz_full.
The following commit adds all the nohz_full cpus to the housekeeping
mask, so that cpu 0 will not get overloaded with timer interrupts if
a large amount of cpus needs to order a timer.

Signed-off-by: Liav Rehana <liavr@mellanox.com>