ARC: [plat-eznps] hookup perf overflow exception to ARC entry code
We define EV_PerfMon for vector index 0x27.
NPS PMU produce exception (NOT interrupt) once counter overflow.

We can have exception while we restore eret at EXCEPTION_EPILOGUE
since we are not in exception mode due to FAKE_RET_FROM_EXCPN.
To avoid eret clobbering we switch back to exception mode. For that,
we are entering a trap and branching back with the special parameter
we give.

Signed-off-by: Noam Camus <noamc@ezchip.com>