ARC: [plat-eznps] Add NPS watchdog.
The NPS has a watchdog mechanism that resets the chip. The watchdog
operates on a group of 6 registers in the CRG block, performs a
countdown, and if not serviced, sends an interrupt and a reset.
The following commit adds the watchdog driver to the NPS, along with
all the needed configurations and device tree changes.

Signed-off-by: Liav Rehana <liavr@mellanox.com>