ARC: [plat-eznps] Set nohz_full cpus as housekeeping.
When using a sleep system call, the cpu needs to order a timer to wake
up. Kernel 4.7 introduced restrict for the cpu that orders the
timer, it needs to be a housekeeping cpu, and that mask is
initialized with all the cpus that are possible in the system and NOT
nohz_full.
The following commit adds all the nohz_full cpus back to the housekeeping
mask, so that cpu 0 will not get overloaded with timer interrupts if
a large amount of cpus needs to order a timer.

Signed-off-by: Liav Rehana <liavr@mellanox.com>
Signed-off-by: Noam Camus <noamca@mellanox.com>